Fix exception checks in unit test and add missed check
Trajectory checks need to also check that the state of
an event is provided as pairs made from tuples or lists.
This sidesteps issues of len((1)) being an attempt to
find the len of an int.

Exception checks in unit test did not actually check
the exception value - need to put outside the with block.